WIP: Move TS glue to separate file
